ww/refactor: db
