+++ database/go.mod
@@ -1,0 +1,5 @@
+module api/internal/database
+
+go 1.21.3
+
+require github.com/joho/godotenv v1.5.1
