+++ src/routes.go
@@ -10,6 +10,9 @@ func NewRouter() *chi.Mux {
 	router := chi.NewRouter()
 
 	router.Get("/healthcheck", handlers.Healthcheck)
+	router.Post("/login", handlers.LoginHandler)
+	router.Post("/image", handlers.ImageHandler)
+
 	// s.router.HandleFunc("/api/", s.handleAPI())
 	// s.router.HandleFunc("/about", s.handleAbout())
 	// s.router.HandleFunc("/", s.handleIndex())
@@ -18,6 +21,5 @@ func NewRouter() *chi.Mux {
 	// router.Get("/healthcheck", healthcheck)
 	// router.Get("/scrape", scrape)
 	// router.Post("/register", registerHandler)
-	// router.Post("/login", loginHandler)
 	return router
 }
